fix(dev-preview): resolve verify failures — shared OAuth type, test narrowing, hook ordering
- Move OAuthLoopbackResult type from electron/services to shared/types/oauth.ts to fix rootDir boundary error
- Narrow result.success in OAuthLoopbackService tests and integration tests
- Restructure BlockedNavBanner hooks to call useCallback before early return (rules-of-hooks)
- Delete untracked tsc artifact files
